# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js Blogg Application API",
    "description": "Yet Another Node.js Blogg Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Contacts",
      "description": "API for contacts in system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/contacts": {
      "post": {
        "summary":"Add the new contact in system",
        "tags": [
          "Contact"
        ],
        "description": "Create new contact in system",
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "description": "Contact that we want to create",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New contact is created",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      },
      "get": {
        "summary":"Get all users from system",
        "tags": [
          "Contacts"
        ],
        "description": "Get all contacts in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of contact that we want to delete",
          "type": "string"
        }
      ],
      "delete": {
        "summary": "Delete user with given ID",
        "tags": [
          "Contacts"
        ],
        "responses": {
          "200": {
            "description": "Contact is deleted",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "required": [
        "first_Name",
        "last_Name",
        "phone"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    }
  }
}